Delete item permissions

+++ backend/src/resolvers/Mutation.js
@@ -29,8 +29,21 @@ const Mutations = {
 	},
 
 	async deleteItem(parent, args, ctx, info) {
+		if (!ctx.request.userId) {
+			throw new Error('Log in first');
+		}
+
 		const where = {id: args.id};
-		const item = await ctx.db.query.item({where}, `{id, title}`);
+		const item = await ctx.db.query.item({where}, `{id title user {id}}`);
+		const ownsItem = item.user.id === ctx.request.userId;
+		const hasPermissions = ctx.request.user.permissions.some(permission => {
+			return ['ADMIN', 'ITEMDELETE'].includes(permission);
+		});
+
+		if (!ownsItem && !hasPermissions) {
+			throw new Error('You don\'t have the necessary permissions to do that')
+		}
+
 		return ctx.db.mutation.deleteItem({where}, info);
 	},
 
@@ -142,7 +155,7 @@ const Mutations = {
 	},
 
 	updatePermissions(parent, args, ctx, info) {
-		if (ctx.request.userID) {
+		if (!ctx.request.userId) {
 			throw new Error('Log in first');
 		}
